+++ SyncTool/DownloadFileRequest.cs
@@ -1,0 +1,2 @@
+﻿namespace SyncTool;
+record DownloadFileRequest(string RelativePath);
