+++ SyncTool/StaticFileConfiguration.cs
@@ -1,0 +1,8 @@
+﻿namespace SyncTool;
+/// <summary>
+/// 静态文件配置
+/// </summary>
+static class StaticFileConfiguration
+{
+    public const string RequestPath = "/File";
+}
