+++ SyncTool/SyncFileInfo.cs
@@ -1,0 +1,2 @@
+﻿namespace SyncTool;
+record SyncFileInfo(string RelativePath, long FileSize, DateTime LastWriteTimeUtc);
